+++ lib/ayadn/set.rb
@@ -15,7 +15,7 @@ module Ayadn
       scroll_config.save
     end
 
-    desc "movie ITEM VALUE", "Set values for #movie (nowwatching)"
+    desc "movie ITEM VALUE", "Set values for movie (nowwatching)"
     map "nowwatching" => :movie
     def movie(*args)
       movie_config = SetMovie.new
@@ -26,6 +26,19 @@ module Ayadn
       end
       movie_config.log(args)
       movie_config.save
+    end
+
+    desc "tvshow ITEM VALUE", "Set values for tvshow (nowwatching)"
+    map "tv" => :tvshow
+    def tvshow(*args)
+      tvshow_config = SetTVShow.new
+      unless args.length != 2
+        tvshow_config.send(args[0], args[1])
+      else
+        abort(Status.error_missing_parameters)
+      end
+      tvshow_config.log(args)
+      tvshow_config.save
     end
 
     desc "nicerank ITEM VALUE", "Set NiceRank filter values"
@@ -178,6 +191,26 @@ module Ayadn
     end
     def hashtag(tag)
       Settings.options[:movie][:hashtag] = tag
+    end
+  end
+
+  class SetTVShow
+    def initialize
+      Settings.load_config
+      Settings.get_token
+      Settings.init_config
+      Logs.create_logger
+    end
+    def log(args)
+      x = "New value for '#{args[0]}' in 'TV Show' => #{args[1]}"
+      puts "\n#{x}\n".color(:cyan)
+      Logs.rec.info x
+    end
+    def save
+      Settings.save_config
+    end
+    def hashtag(tag)
+      Settings.options[:tvshow][:hashtag] = tag
     end
   end