+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# 1.6.0 - 'Baskerville'
+# 1.6.0 - 'Galadriel'
 
-- New command: 'movie' (nowwatching). Create a post from a movie title (with link + movie poster). Customisable hashtag (default: '#movie').
+- New command: 'movie'. Create a post from a movie title (with link + movie poster). Customisable hashtag (default: '#nowwatching').
+- New command: 'tvshow'. Create a post from a TV show title (with link + show poster). Customisable hashtag (default: '#nowwatching').
 - New: see the target of your PM when writing a message.
 - New: delete several posts at once.
 - New: delete several PMs at once.